+++ src/main/java/gov/usgs/wma/waterdata/collections/CollectionsController.java
@@ -1,5 +1,8 @@
 package gov.usgs.wma.waterdata.collections;
 
+import static gov.usgs.wma.waterdata.collections.CollectionParams.PARAM_COLLECTION_ID;
+import static gov.usgs.wma.waterdata.collections.CollectionParams.PARAM_FEATURE_ID;
+
 import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
@@ -25,7 +28,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 @Tag(name = "Observations - OGC api", description = "Feature Collections")
 @RestController
 public class CollectionsController {
-	
 	protected CollectionsDao collectionsDao;
 
 	protected CollectionParams collectionsParams;
@@ -70,7 +72,7 @@ public class CollectionsController {
 		)
 	@GetMapping(value = "collections/{collectionId}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public String getOgcCollection(@RequestParam(value = "f", required = false, defaultValue = "json") String mimeType,
-			@PathVariable(value = "collectionId") String collectionId, HttpServletResponse response) {
+			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId, HttpServletResponse response) {
 
 		
 		return resultOr404(response, 
@@ -97,7 +99,7 @@ public class CollectionsController {
 			@RequestParam(value = "limit", required = false, defaultValue = "10000") int limit,
 			@RequestParam(value = "startIndex", required = false, defaultValue = "0") int startIndex,
 			@RequestParam(value = "bbox", required = false) List<String> bbox,
-			@PathVariable(value = "collectionId") String collectionId, HttpServletResponse response) {
+			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId, HttpServletResponse response) {
 
 		int count = collectionsDao.getCollectionFeatureCount(collectionsParams.buildParams(collectionId));
 
@@ -132,8 +134,8 @@ public class CollectionsController {
 	@GetMapping(value = "collections/{collectionId}/items/{featureId}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public String getOgcCollectionFeature(
 			@RequestParam(value = "f", required = false, defaultValue = "json") String mimeType,
-			@PathVariable(value = "collectionId") String collectionId,
-			@PathVariable(value = "featureId") String featureId, HttpServletResponse response) {
+			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId,
+			@PathVariable(value = PARAM_FEATURE_ID) String featureId, HttpServletResponse response) {
 
 		return resultOr404(response, 
 				collectionsDao.getCollectionFeatureJson(collectionsParams.buildParams(collectionId, featureId)));