+++ src/main/java/gov/usgs/wma/waterdata/collections/CollectionsStatTimeSeriesController.java
@@ -2,7 +2,6 @@ package gov.usgs.wma.waterdata.collections;
 
 import static gov.usgs.wma.waterdata.collections.CollectionParams.PARAM_COLLECTION_ID;
 import static gov.usgs.wma.waterdata.collections.CollectionParams.PARAM_FEATURE_ID;
-import static gov.usgs.wma.waterdata.collections.CollectionParams.PARAM_TIME_SERIES_ID;
 
 import javax.servlet.http.HttpServletResponse;
 
@@ -13,8 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import gov.usgs.wma.waterdata.timeseries.TimeSeriesDao;
-import gov.usgs.wma.waterdata.timeseries.geojson.TimeSeriesGeoJSON;
+import gov.usgs.wma.waterdata.collections.observations.StatisticalFeatureGeoJSON;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -24,42 +22,42 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "Statistical Observations Datasets", description = "Feature Statistical Time Series Observations, such as min, max, or median.")
 @RestController
-public class ObservationsStatTimeSeriesController {
-	
-	protected static final String URL_STATISTICAL_TIME_SERIES 
-		= "collections/{"+PARAM_COLLECTION_ID+"}/items/{"+PARAM_FEATURE_ID+"}/observations/statistical-time-series/{"+PARAM_TIME_SERIES_ID+"}";
-	
-	protected TimeSeriesDao timeSeriesDao;
+public class CollectionsStatTimeSeriesController {
+
+	protected static final String URL_STATISTICAL_TIME_SERIES_COLLECTION
+		= "collections/{"+PARAM_COLLECTION_ID+"}/items/{"+PARAM_FEATURE_ID+"}/observations/statistical-time-series";
+
+	protected CollectionsDao collectionsDao;
+
 
 	@Autowired
-	public ObservationsStatTimeSeriesController(TimeSeriesDao timeSeriesDao) {
-		this.timeSeriesDao = timeSeriesDao;
+	public CollectionsStatTimeSeriesController(CollectionsDao collectionsDao) {
+		this.collectionsDao = collectionsDao;
 	}
 
 	@Operation(
-			description="Return GeoJSON Data specific to the requested Monitoring Location and Time Series.",
+			description="Return JSON Data list of statistical time series available for the requested Monitoring Location.",
 			responses= {
 					@ApiResponse(
 							responseCode="200",
-							description="GeoJSON representation of the Time Series.",
-							content=@Content(schema=@Schema(implementation=TimeSeriesGeoJSON.class))),
+							description="GeoJSON representation of the Statistical Time Series list.",
+							content=@Content(schema=@Schema(implementation=StatisticalFeatureGeoJSON.class))),
 					@ApiResponse(
 							responseCode="404",
-							description="The requested Time Series was not found.",
+							description="The requested collection and feature combination was not found.",
 							content=@Content())
 			},
 			externalDocs=@ExternalDocumentation(url="https://github.com/opengeospatial/omsf-profile/tree/master/omsf-json")
 		)
-	@GetMapping(value=URL_STATISTICAL_TIME_SERIES, produces=MediaType.APPLICATION_JSON_VALUE)
+	@GetMapping(value=URL_STATISTICAL_TIME_SERIES_COLLECTION, produces=MediaType.APPLICATION_JSON_VALUE)
 	public String getTimeSeries(
 			@PathVariable(value=PARAM_COLLECTION_ID) String collectionId, // ex: networkId,
 			@PathVariable(value=PARAM_FEATURE_ID) String featureId, // ex: monitoringLocationId
-			@PathVariable(value=PARAM_TIME_SERIES_ID) String timeSeriesId, //ex: USGS-123456
 			HttpServletResponse response) {
-		
+
 		// verify the collection and feature exist before fetching the time series
-		String json = timeSeriesDao.getTimeSeries(collectionId, featureId, timeSeriesId);
-		
+		String json = collectionsDao.getStatisticalTimeSeries(collectionId, featureId);
+
 		if (null == json) {
 			response.setStatus(HttpStatus.NOT_FOUND.value());
 		}