+++ CHANGELOG.md
@@ -6,10 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/USGS/time-series-services/compare/time-series-services/compare/0.4.0...master)
 
-## [0.4.0](https://github.com/USGS/time-series-services/compare/time-series-services/compare/0.2.0...0.4.0)
+## [0.4.0](https://github.com/USGS/time-series-services/compare/time-series-services-0.2.0...time-series-services-0.4.0)
 ### Added
 -   Time Series GeoJSON endpoint.
--   Time Series JSON endpoint within collections
+-   Feature Statistical Time Series collection
+-   Feature Statistical Time Series observations
 -   Collections GeoJSON endpoint.
 
 ## [0.2.0](https://github.com/usgs/time-series-services/compare/time-series-services-0.1.0...time-series-services-0.2.0)